pyproject.toml
[tool.poetry]
name = "auto-label-pipeline"
version = "0.1.0"
description = "\"Automatically refine labels using embeddings, classifiers and CleanLab\""
authors = ["todd-cook <665389+todd-cook@users.noreply.github.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "auto_label_pipeline"}]

[tool.poetry.dependencies]
python = "^3.10"
dvc = "^2.55.0"
PyYAML = "^6.0"
cleanlab = "^2.3.1"
nltk = "^3.8.1"
numpy = "^1.24.3"
pandas = "^2.0.1"
scikit-learn = "^1.2.2"
fasttext-wheel = "^0.9.2"
types-pyyaml = "^6.0.12.9"
types-requests = "^2.28.11.17"
tqdm = "^4.65.0"
isort = "^5.12.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
codecov = "^2.1.13"
coverage = "^7.2.3"
mypy = "^1.2.0"
pigar = "^2.0.6"
pylint = "^2.17.3"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-mypy = "^0.10.3"
pytest-pylint = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
